requirements.txt
pydantic>=2.0
fastapi>=0.110
uvicorn
httpx
asyncpg
structlog
azure-identity
azure-keyvault-secrets
pytest
pytest-asyncio
hypothesis
coverage
python-dotenv
pgvector
sqlalchemy[asyncio]
alembic
respx
pytest-recording
vcrpy
testcontainers
